+++ message/messages.go
@@ -38,7 +38,7 @@ const (
 	CIDRNetmaskOutOfRange = "The value of the netmask should be between {{ min }} and {{ max }}."
 	InvalidIP             = "This is not a valid IP address."
 	InvalidJSON           = "This value should be valid JSON."
-	InvalidLuhn           = "Invalid card number."
+	InvalidLUHN           = "Invalid card number."
 	InvalidTime           = "This value is not a valid time."
 	InvalidULID           = "This is not a valid ULID."
 	InvalidUPCA           = "This value is not a valid UPC-A."